Fix positive scaling enforcement
Old line:    "if (scaling<10-6); scaling = 10-6; end;"
Fairly certain it should be 10e-6. Otherwise the scaling param is broken and sets to the same for all subjects. Number of iterations of crossvalidated minimization is now fewer, and the overall time to fit all models is much shorter. Results and iterations are in line with previous versions of function on lab server.