Fixed Model handling
Adjusted line 25 to handle fixed models with unique fixed G for each subject.
Since pcm_calculateG is only ever called for overall fitting (never for crossval fitting) for fixed models, we can average across all subjects' Gs in this fixed model. 
For example, let's say we submit a fixed model structure, M. Under this fixed model, we have a unique predicted G for each subject  that we do not want to iterate to achieve the best fit (e.g. do a subject's (co)variances between finger patterns scale as a function of the number of finger presses?). In this structure, M.Gc is a 3 dimensional matrix, where the third dimension contains a G for each subject.