Fixed Model Handling
Adjusted estimation of crossvalidated G for fixed models to handle cases where fixed model has unique G for each subject.
 
For example, let's say we submit a fixed model structure, M. Under this fixed model, we have a unique predicted G for each subject  that we do not want to iterate to achieve the best fit (e.g. do a subject's (co)variances between finger patterns scale as a function of the number of finger presses?). In this structure, M.Gc is a 3 dimensional matrix, where the third dimension contains a G for each subject.